BOL-375 refactor: setting model data class file 분리
