+++ data/src/main/java/com/yapp/bol/data/model/setting/SettingTermDTO.kt
@@ -1,0 +1,8 @@
+package com.yapp.bol.data.model.setting
+
+data class SettingTermDTO(
+    val code: String,
+    val title: String,
+    val url: String,
+    val isRequired: Boolean,
+)
